data_types size in byte(s)

+++ 0x00-hello_world/6-size.c
@@ -1,14 +1,14 @@
 #include <stdio.h>
 /**
-  * Prints size of various data types
+  * main - Prints size of various data types
   * Prints type of comp compiled on 
-  * Always return 0
+  * Return: Always 0.
   */
 int main(void)
 {
 	printf("Size of a char: %zu bytes\n",sizeof(char));
 	printf("Size of int: %zu bytes\n",sizeof(int));
-	printf("Size of long int: %zu bytes\n",sizeof(int));
+	printf("Size of long int: %zu bytes\n",sizeof(long int));
 	printf("Size of long long int: %zu bytes\n",sizeof(long long int));
 	printf("Size of float %zu bytes\n",sizeof(float));
 	return (0);
